Title: X 달리기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250204/X 달리기/x-run.cpp
@@ -8,6 +8,6 @@ int main() {
     
     cin >> x;
 
-    cout << 2 * sqrt(x - 1);
+    cout << int(2 * sqrt(x - 1));
     return 0;
 }
